media-libs/opencv: pass CMAKE_CUDA_ARCHITECTURES for FindOpenMP
cmake-3.31.0 extended FindOpenMP.cmake to look up CUDA flags
and detect compute capabilites if the CUDA language support
is enabled. This requires write access to /dev/nvidiactl.
Iff the portage user is not in the video group this will fail.
To avoid this we set the lowest supported architecture by CUDA-11.8 and
12.x, sm_50. This can be overridden by setting CUDAARCHS.

Closes: https://bugs.gentoo.org/955860
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>