x11-drivers/nvidia-drivers: describe subslot in metadata.xml
In general packages aren't even supposed to depend on this except at
most with [static-libs], and if they do then the binding operator
does not mean much. The static library also hasn't really changed in
ages and is effectively deprecated and could be removed anytime (thus
rebuilding for it shouldn't be meaningful).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>