wine.eclass: limit -std=gnu17 to <wine-10
chiitoo mentioned that (at least) 10.7 works without this, and
that the last needed fixes were in 10.0-rc1, so I tried 10.0
and it was also fine.

vanilla-9.0.1 does not seem to have received a hotfix for this and
still fails, there's also proton-9.0.4 to worry about so will just
keep passing it for <10 until they're gone.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>