dev-java/*: restict javacup, jflex and qdox:1.12 to revisions
There is a circular dependency between javacup and jflex and in the past
we used to build jflex with a bundled javacup-runtime. The newer version
of jflex-1.9.1 however does not provide such bundled javacup-runtime.

This commit restricts the existing revisions to what we have in order to
allow updating jflex.

- javacup-11b_p20160615-r1
  Restricted to ~jflex-1.6.1 in order to allow switching the emerge
  order. javacup-11b_p20160615-r2 will be used by jflex-1.9.1 but not
  depend on it.

- jflex-1.6.1-r4
  Restricted to max javacup-11b_p20160615-r1. The next revision,
  javacup-11b_p20160615-r2, will not depend on jflex.

- qdox-1.12.1-r7
  The patch 'jflex-1.6.1.patch' which is used by this revision does not
  allow building with a newer version of jflex.
  We restict this revision to ~dev-java/jflex-1.6.1 and will add another
  revision which can build with the newer jflex.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>
Part-of: https://github.com/gentoo/gentoo/pull/44017
Signed-off-by: Sam James <sam@gentoo.org>