Revert "rust.eclass: revert simplified dependency simplification"
This reverts commit 34b74faa06a90bf9d4f62ecfca746b380d60517a, i.e.
reapplies the simplified dependency specification.

We had to revert this before because of the collision issues when
switching to slotted Rust where the same Rust version existed before
in a non-slotted variant. We have a sanity check and workaround for that
in pkg_postinst for affected versions now, so there shouldn't be a reason
we can't do this.

This may ease depcleaning older Rust in some cases w/o having to rebuild
packages against new Rust.

Bug: https://bugs.gentoo.org/943143
Bug: https://bugs.gentoo.org/943206
Bug: https://bugs.gentoo.org/953884
Signed-off-by: Sam James <sam@gentoo.org>