dev-python/twisted: enable py3.14
Tested multiple rev-deps, and they work in py3.14 compat.

tests don't pass for any python target for me, but the failures & errors
were the same for me for py3.13 and py3.14 (which were different for
py3.12 - WTF).

so I think py3.14 is as ready as py3.13, so I see no issue adding it
also to PYTHON_TESTED.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>