python-utils-r1.eclass: python_optimize, use sitedir, not sys.path
Default to optimizing the `site-packages` directory rather than
the whole `sys.path` when no path is passed to `python_optimize`.
Technically, this is a breaking change, but it is unlikely that any
ebuilds have been actually relying on the old behavior, and in the worst
case our QA checks will note the missing .pyc files.

The old logic was assuming that a custom package may install overrides
to `sys.path` that we would need to account for.  However, this assumes
that said overrides would actually start working *before* the package is
installed (creating hard-to-reproduce bugs) -- and in these cases, we
really prefer listing special paths explicitly anyway.

Signed-off-by: Michał Górny <mgorny@gentoo.org>
Part-of: https://github.com/gentoo/gentoo/pull/41560
Signed-off-by: Michał Górny <mgorny@gentoo.org>