java-utils-2.eclass: re-enable PORTAGE_QUIET
There are packages like javacup and jflex which cannot be built from
source without using a pre-built runtime version of itself. Re-emerging
these packages using the installed instead of the bundled pre-built
version was causing the java-pkg_getjar() function to trigger
java-pkg_ensure-dep() to issue a "QA Notice: java-pkg_ensure-dep:"
message.

This commit ports the PORTAGE_QUIET variable to java-pkg_getjar() so
that it works again as it did in the past for ant-based ebuilds.

Bug: https://bugs.gentoo.org/937047
Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>