linux-info.eclass: Set the correct ARCH for make in getfilevar
Currently, it unsets ARCH, so it uses the build architecture. This is
fine if you have the full kernel sources, which is usually the case.
However, an install-extmod-build script was added to Linux fairly
recently that installs just the files needed to build out-of-tree
modules in a distro-neutral manner. It saves space by only including the
arch files for the target architecture.

I checked the history, and tc-arch-kernel didn't exist when the unset
line was added. If it had, this approach probably would have been taken
at the time.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>