media-libs/oidn: Fix hip compilation errors
Debundling the included sci-libs/composable-kernel lib lead to
compilation issues. Because it took over 12 hours to compile
sci-libs/composable-kernel on my computer, and less than 10 minutes to
compile oidn with the bundled composable-kernel lib, I think it is best
to not debundle this library.

Closes: https://bugs.gentoo.org/955869
Signed-off-by: Sebastian Parborg <darkdefende@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/42204
Signed-off-by: Sam James <sam@gentoo.org>