dev-util/nvidia-cuda-toolkit: add clang useflag to force specific compiler
The at-least-one-of dependency on clang or gcc allowed the PM to depclean one of
them. If e.g. clang is installed via LLVM_SLOT it allows older gcc to be
depcleaned even if it is the system compiler. Causing the hostcompiler lookup to
fail.

Bug: https://bugs.gentoo.org/956046
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/42202
Signed-off-by: Sam James <sam@gentoo.org>