linux-mod-r1.eclass: Fix building with different arch than KV_OUT_DIR
The kernel we're building against may have had its tools (e.g. modpost)
built for the target arch or even some other arch than the we're
building with now. To work around this, rebuild those tools with make
modules_prepare when necessary.

Doing this in pkg_setup is not ideal, but some ebuilds have their own
src_configure and src_compile functions. It will not trigger in the vast
majority of cases, and it is not done when installing binary packages.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>