dev-qt/qtbase: add wall of blockers for old <dev-qt/*-${PV} in live
This is the same as commit 2455f4d30364e65eccbcfee4d8edd6bd8742cba2
that was done for dev-qt/qtcore:5. Been found to still be a problem if
e.g. qtvirtualkeyboard-6.8.2 is installed while using Qt 6.8.3+ where
it causes issues for sddm/breeze theme when it tries to load the
incompatible qtvirtualkeyboard plugin. This can happen when a user
have it installed while in a depcleanable state but does not depclean,
and the package manager is allowed to not update in this stuation.

Does have the downside of causing more confusing portage output when
users upgrade, but hasn't been known to cause actual update problems
(issues will most likely be due to something else than these blockers).

Not *all* these packages can cause problems (some are unlikely to be
dlopen'ed), but it's simpler to just list all packages with libraries.
List was copied from stabilization-groups' qt6.group given it has all
packages w/ Qt6, excluding qtbase (no need to self-block), pyside,
qttranslations, and qt-docs.

fwiw should also block higher versions in case of Qt downgrade, but
that is an unusual occurance combined with unusual lack of depclean
and this does not need to handle everything (would need to make +1
blockers to allow revbumps and kind of meh to keep in the ebuild).

Qt5 blocks <QT5_PV aka $(ver_cut 1-3), but a similar variable has not
been used with Qt6 yet and blocking _rc as well does not hurt even if
they will not "mismatch version" with non-_rc, so just using PV.

For now only adding in live, and it will start being used with 6.9.1
and can see to any issue then.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>