qt6-build.eclass: use dot-a if QT6_HAS_STATIC_LIBS=1
Only relevant with USE=custom-cflags at the moment until we
potentially remove the filter-lto.

Considered QT_FEATURE_ltcg so QtModuleHelpers.cmake handles adding
-ffat-lto-objects *only* for static libs (less wasteful) but
1. FEATURE_ltcg is typo'ed there (lacks QT_), 2. it is only
activated with GCC (not clang), and 3. not 100% sure -ffat is passed
everywhere that it needs to be that way. May revisit.

Also considered adding this to ebuilds directly, but there is 5
that need this and it needs defining 1 or even 2 phases in some.

(not putting this behind USE=custom-cflags so that it will not need
adjusting if we remove filter-lto later)

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>