dev-python/wxpython: drop sip-6.11 upper bound
Unfortunately needs a revbump to clear up VDB of the upper bound and
let portage update sip with default --with-bdeps=y. Albeit likely not
many users had time to build with it.

fwiw only do it for latest, in process of stabilization and no sense in
making stable 4.2.2 users rebuild before an update.

Also block 6.11.1-r0 in latest for now to ensure it's updated, but the
jarring || ( ( > < ) > ) won't need to be kept long term nor need a
revbump if it's revmoed.

Bug: https://bugs.gentoo.org/956566
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>