dev-util/coccinelle: Add 1.3.0
There was new release [1] which compared to what we currently
have in portage introduces plenty of new features (many of which
are aimed at supporting newer C/C++ standards), though some were
introduced in the 1.2.0 release.

The patches applied are either taken from 1.1.1 version (when
they apply cleanly) or rebased
(coccinelle-1.3.0-find-xxdate.patch) where the line removed
needed to be changed slightly.

Then there's one more patch that's new:
coccinelle-1.3.0-python-3.13.patch which is a patch for pyml
(ocaml bindings to python) so that python-3.13 is supported. It's
backport of the following upstream commit [2]. Unfortunately,
pyml is not in portage yet, so coccinelle uses the bundled
version which hasn't been updated in ages.

1: https://coccinelle.gitlabpages.inria.fr/website/distrib/changes.html
2: https://github.com/ocamllibs/pyml/commit/68bf34b79bd43d145fbf33f8bbd7c4eb520cae34
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/42276
Closes: https://github.com/gentoo/gentoo/pull/42276
Signed-off-by: Sam James <sam@gentoo.org>