llvm-runtimes/libgcc: Allow cross-compilation
CMake compiler tests compile and run a test artifact. That works well on
host targets, but doesn't work with cross wrappers, since the produced
binary is of a foreign target, so executing it without an user-space
emulator is impossible.

Given that, let's just skip the compiler tests for cross targets.

Bug: https://bugs.gentoo.org/956749
Signed-off-by: Michal Rostecki <vadorovsky@disroot.org>