x11-drivers/nvidia-drivers: warn about intel+nvidia laptops in 0/575
Some users may have already upgraded and missed the message but it
does not hurt to add one anyway.

Not heard of this from Gentoo users yet (the beta was unkeyworded),
but it seems(?) to be a widespread issue on other distros with 575.
Supposedly does not manifest with 570.x, and unclear if kernel 6.15.x
is related.

Thankfully, USE=kernel-open is off-by-default, 575.x is not a stable
candidate either way, and normal desktops seem fine, so it should not
affect *that* many users. Do not believe this issue requires masking
this version (lot of interesting fixes for other users).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>