{dev-lang,sys-devel}/rust*: add 1.88.0_beta20250526
This is our first beta for 1.88.0, corresponding to beta.4 upstream.
It is also the first release where `beta` and `live` ebuilds get their
patches from gitweb rather than ${FILESDIR}.

Restrictions on building beta with only the last stable have been
dropped. Will assess over the remainder of 1.88 and 1.89 betas if this
change can remain or if we have to revert.

Signed-off-by: Matt Jolly <kangie@gentoo.org>