dev-lang/rust: drop build `x.py` verbosity
After discussion with upstream, reduce the number of `v`s passed to
of our `x.py` invocations. There is no documentation on why `-vvv` was
added but it's suspected to be development troubleshooting shrapnel.

`-vv` is retained for tests where the extra verbosity may actually be
useful, however for most cases `-v` is suitable as it includes the
invocation along with any warnings or errors; we probably don't need
thousands of backtraces or diagnostic messages.

Signed-off-by: Matt Jolly <kangie@gentoo.org>