python-utils-r1.eclass: Fix epytest with >=pytest-8.4.0
Remove the `-Werror::PytestUnhandledCoroutineWarning` argument
from `epytest`, since the relevant warning is no longer present
in `>=dev-python/pytest-8.4.0`.  The newer versions throw an exception
instead, so we don't need any argument.

We could technically still pass it conditionally to pytest version --
but that breaks dev-python/pytest's test suite on upgrade, and working
around that does not seem worth the effort.  In the end, this was a QA
check to catch pathological cases, and it should be sufficient to catch
them with the latest pytest version.

Signed-off-by: Michał Górny <mgorny@gentoo.org>