x11-drivers/nvidia-drivers: blacklist nova_core as well
This is new rust-written open-source drivers for nvidia newly
made available in kernel 6.15, albeit quite work-in-progress and
not quite sure if they are even usable yet. And, like nouveau,
they would conflict if used.

Not worth revbumps given users would've had to go quite out of their
way to enable it manually and should know what they are getting
into. It will either way propagate on kernel bumps due to requiring
to rebuild nvidia-drivers.

Blacklisting given "possible" that dist-kernels will enable it by
default in the future albeit it is quite experimental at this stage
and needs rust, so I doubt it'll happen with 6.15.x.

Note nvidia-drm.fbdev=1 may or may not be able to unload it which
*could* allow unblacklisting (same for nouveau), but this feels kind
of fragile and we have it off by default for now. It is more typical
to use fbdev=1 to unload simpledrm after using for early boot display.

Technically not meaningful in nvidia-390/470.conf given nova is not
usable for these old cards, but no real reason not to keep in sync.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>