kernel-build.eclass: fix signing non-zboot image for secureboot
If the kernel has been built without EFI_ZBOOT support (requires 6.1+)
then the resulting kernel image will be an Image.gz that we cannot
sign with sbsign (on arm64 and riscv). So, uncompress this image,
then add the signature, and finally recompress it with the same
options that the kernel Makefiles use.

Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>