net-libs/webkit-gtk: add 2.48.3
Updates:

 * Fix the riscv build, re-keyword for ~riscv
 * Drop the obsolete ICU patch
 * Disable the new drag-drop, speech synthesis, and touch support via
   mycmakeargs (path of least resistance, does not involve any new flags
   or dependencies).

These are usable in Evolution and with the MiniBrowser at least. There
are some caveats however:

 * Does not build for me with USE="-jumbo-build"
 * Does not build for me with clang LTO
 * The -r600 doesn't seem to work with epiphany (all pages hang)
 * The masked 2.48.3 (sans -rXXX) was not tested

The ~riscv keywords are the only reason we would need to keep the old
vulnerable 2.44.x ebuilds in the tree, and the "current" versions can
fail to build too, so I think this is a win despite the setbacks. The
old versions are still there if they cause less problems for someone.

Bug: https://bugs.gentoo.org/942369
Closes: https://bugs.gentoo.org/951085
Closes: https://bugs.gentoo.org/955769
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>