dev-python/google-auth: drop dev-python/moto from test deps
moto isn't used at all in this package, nor was it ever used. Searching
git history for the word "moto" finds nothing. No mention of moto in
github issues/pull requests/...

It was added as fix for [1], but there was no "moto" in the build.log

We can conclude it was added by mistake and was never needed.

[1] https://bugs.gentoo.org/723554

Also-checked-by: Michał Górny <mgorny@gentoo.org>
Also-checked-by: Sam James <sam@gentoo.org>
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>