dev-qt/qtwayland: update live
Most code was moved to qtbase alongside their tests, and
this package is now primarily for the compositor bits.

Considered still keeping USE=compositor but "wayland_server"
feature was moved to qtbase as well and would need control it
*there* and depend on qtbase[compositor=] which provides rather
poor UX due to portage giving misleading warnings (not as bad
for e.g. opengl=,vulkan= because they are set globally by
profiles). On the bright side, many users may be able to
depclean qtwayland altogether so it won't be wasteful.

Formerly skipped tst_compositor is still here but let's
tentatively allow it again (seems to pass).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>