Revert "dev-qt/qtdeclarative: set CMAKE_QA_COMPAT_SKIP=1"
On second thought, will set this in the eclass. Affects at least one
more package (qtmultimedia), and unused or replaced CMakeLists.txt is
a common trend in Qt and this may be volatile. If something breaks
with cmake-4, would rather that it just breaks rather than do the =3.5
workaround and will fix it then.

This reverts commit 51c2217d6c13c8009c08b3b9de46ff3491e792be.

Bug: https://bugs.gentoo.org/957475
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>