net-irc/quassel: drop dependency dev-libs/libdbusmenu-qt
The last upstream snapshot "release" of libdbusmenu-qt happened in 2016,
with no development since then. It depends on Qt 5, which is scheduled
for removal (bug #948836). Quassel and KNotifications are the last
 consumers of the library.

Libdbusmenu was introduced upstream as an optional dependency in commit
5d9af8ed007d38faf3995ea18174249121f246fe in 2010, to improve system
integration of the menu and its respective icons in the system tray.
(As a side note, even upstream could not test the feature themselves
when it was introduced.)

If upstream becomes active again and wants to resurrect the feature with
a migration to Qt 6, it would need to be re-implemented or dropped
anyway.

Bug: https://bugs.gentoo.org/938826
Signed-off-by: Johannes Huber <johu@gmx.de>