dev-util/ctags: rename versions
Move away from the strange/unconventional snapshot scheme introduced
in 6635a8c6cb2423209613660b53d8360ebe621a75 towards the traditional
scheme we use for snapshots (LAST_VERSION_pXYZ). Prompted by upstream
tagging 6.2.0 so we need this for things sort correctly.

I thought I'd remarked on this in a commit message before but decided
not to fix it up at the time (it was pre-existing) but I can't find that.

Signed-off-by: Sam James <sam@gentoo.org>