sys-devel/patch: fix incorrect reliance on split-usr
It hardcodes a compile-time path to `ed`. This is known and handled
already because ${ED} is part of portage and does not point to `ed`. But
the "fix" was to do a $PATH lookup, which in turn broke binpackage
hosting (and profile migration, likely).

The package doesn't have USE=split-usr and we shouldn't add one just for
this. Make sure the portable API filename is used.

Bug: https://bugs.gentoo.org/470210
Closes: https://bugs.gentoo.org/957593
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>