www-servers/tomcat: tomcat-r2.init, fix classpath
Tomcat depends on ant.jar which is one of two jars of dev-java/ant and
so far it was built with $(java-pkg_getjar --build-only ant ant.jar).
Using '--build-only' here was causing java-dep-check to complain with
'Class files not found via DEPEND in package.env' for several of ant's
classes.

Solving these complaints by building without --build-only for ant leads
to having 'ant.jar@ant' in the DEPEND line of package.env which was
causing a java-config call of the init script to fail with 'ERROR:
Package ant.jar@ant was not found!'.

This commits adds a hack which removes the ant.jar@ part from the init
script's local DEPEND variable.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>