x11-misc/xkeyboard-config: workaround portage collision issue
Don't really like this, but unless portage either relax its
checks or improve merging process there's probably(?) not a
better option (would still need workaround for old portage
anyway).

Can probably drop this sometime in 2+ years after 2.45 is old
enough, or if the symlink is ever removed.

Unsure if maintainer would rather the workaround done a bit
differently but that can be changed later (just do not want
to leave this broken for longer).

Closes: https://bugs.gentoo.org/957712
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>