x11-misc/xkeyboard-config: fix collision workaround for binpkgs
Was hoping to "do nothing at all" when not needed, but cannot test
the current system in src_install for binpkgs. fwiw, we can still
avoid touching ${EROOT} for nothing unless xkb is a directory
(still with -f in case users did anything weird with permissions).

Also revbump just in case binpkgs were made already either with or
without the previous version of the workaround.

Closes: https://bugs.gentoo.org/957712
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>