x11-misc/gigolo: avoid automagic deps on gtk+[X]
e.g. the upstream code checks whether GDK_WINDOWING_X11 is defined
by the gtk headers, and if so will compile against the x11 symbols
it provides. This means that gigolo built on a system with gtk+[-X]
will be compatible with anything, but when built on a system with
gtk+[X], requires that at runtime.

Use a Gentoo-specific macro hack to hide the GTK implementations when we
don't want it, alongside new USE flags to control the dependency usage.

Closes: https://bugs.gentoo.org/957781
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>