media-gfx/geeqie: avoid automagic deps on gtk+[X]
e.g. the upstream code checks whether GDK_WINDOWING_X11 is defined
by the gtk headers, and if so will compile against the x11 symbols
it provides. This means that geeqie built on a system with gtk+[-X]
will be compatible with anything, but when built on a system with
gtk+[X], requires that at runtime.

Use a Gentoo-specific macro hack to hide the GTK implementations when we
don't want it, alongside new USE flags to control the dependency usage.

In this case, upstream supports wayland fine, but doesn't need any
backend-specific code for it. There are some X11-specific hacks.

Bug: https://bugs.gentoo.org/957840
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>