profiles: package.use.force x11-libs/fltk[static-libs]
Not ideal, but would rather not deal with this downstream, esp. if
going to modify exports which becomes bit of a Gentoo invention
(the *FLAGS bit from fltk-config that we need to do is already meh).

Could remove the USE entirely + enable, but use.force is a way to
show awareness of the issue (notably for QA checks), and still
allows to unforce for users that really do not want it.

Not overly necessary to force for old version too (pre-cmake), but
do it anyway for fltk-config.

At least, static-libs does not add notable space usage here.

If do not like this, please take it upstream and will update on bump.

Bug: https://bugs.gentoo.org/940572
Closes: https://github.com/gentoo/gentoo/pull/42633
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>