licenses: add NVIDIA-2023 and NVIDIA-2025
When diffing for the nvidia-drivers-570.169 bump noticed a 2023->2025
timestamp change in the license alongside other bits and well, we did
not have the 2023 version either which is used by all current
non-masked/legacy (>=535) versions in the tree.

Considered doing -r3 and -r4 but given nvidia is giving a date in the
license now may as well use the year to be easier to track.

Reading it, certainly still not a free license but there still does
not seem to be anything preventing bindist and mirror (in my eyes
at least), and it is no longer a Linux+FreeBSD-only exception.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>