dev-qt/qtgraphs: new package, add 6.9.1 + live
Not needed by anything in ::gentoo "yet", but qtcharts is
being deprecated in Qt 6.10 and projects are told to migrate
to qtgraphs, so let's get this ready.

Unknown at the moment how much quick3d will be needed/wanted
(required for QtGraphsWidgets), left off-by-default for now
but that can be revisited if packages start needing it.

quick3d is kept conditional given it can be flaky on some
arches (may need to mask depending on future keywords) and
is semi-big to build.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>