net-misc/sunshine: Version bump to 2025.122.141614 and refresh 9999
Sorry this took so long, and thank you to the users who contributed fixes. This
has always been a tricky package, but hopefully we won't have to go through this
amount of change again.

Unfortunately, Boost 1.87 is still required for now. I tried to port it to 1.88,
but Sunshine uses parts of boost::process that just aren't there anymore. I'll
send what I have so far upstream shortly.

Bug: https://bugs.gentoo.org/955560
Bug: https://bugs.gentoo.org/956631
Closes: https://bugs.gentoo.org/947046
Closes: https://bugs.gentoo.org/951973
Closes: https://bugs.gentoo.org/956006
Closes: https://github.com/gentoo/gentoo/pull/42477
Signed-off-by: James Le Cuirot <chewi@gentoo.org>