sys-libs/libblockdev: patch CVE-2025-6019
Quoting myself on the bug:
> This is from https://www.openwall.com/lists/oss-security/2025/06/17/4 but
> it's a little complicated in that it relies on a privilege escalation bug
> from unprivileged -> polkit 'allow_active' (bug 958338) that we aren't
> affected by.
>
> But nonetheless, supposing such another such LPE exists, this could form
> part of a chain from 'allow_active' -> root, so it still matters.

My intention is not to rush stabling 3.3.0 as just bumped to it (that
commit bumping to 3.3.0 lands in the same push as this) and we were
a bit behind before so want to give a little bit of time for any regressions
to be reported. Combined with the above, we don't need to hurry s.t. we do
it with no time in ~arch at all.

Bug: https://bugs.gentoo.org/958338
Bug: https://bugs.gentoo.org/958339
Signed-off-by: Sam James <sam@gentoo.org>