www-apps/rutorrent: clean up useless files
The archives contain metadata from version control and an editor. We
always deleted that but now adding ".vscode" and moving things to
src_prepare. Same goes for the "tests" directory which is not for
end-users.
We now move documentation out of the webapp with dodoc.

Signed-off-by: Henning Schild <henning@hennsch.de>
Part-of: https://github.com/gentoo/gentoo/pull/42652
Closes: https://github.com/gentoo/gentoo/pull/42652
Signed-off-by: Viorel Munteanu <ceamac@gentoo.org>