llvm-runtimes/libcxx{abi,}: Allow to build without clang configuration
When the full LLVM toolchain is upgraded (e.g. from 19 to 20), the order
of ebuild upgrades is:

- llvm-core/clang
- [...]
- llvm-runtime/libcxxabi
- llvm-runtime/libcxx
- llvm-core/clang-runtime

libcxxabi, libcxx are emerged with a new clang, but since clang-runtime
is is not yet upgraded, that new clang doesn't have correct
configuration. On the other hand. clang-runtime wouldn't be complete
without up-to-date libcxx.  Therefore, we need to tell libcxx{abi,}
what runtime libraries should be used.

Bug: https://bugs.gentoo.org/951445
Signed-off-by: Michal Rostecki <vadorovsky@disroot.org>