python-utils-r1.eclass: Introduce EPYTEST_PLUGINS
Introduce a new EPYTEST_PLUGINS variable that can be used to control
which plugins are used by epytest.  It provides a convenient method
to disable plugin autoloading, as well as specify which plugins to load.
It accepts package names, and includes a mapping from package names
to `-p` argumetns for a number of plugins currently in ::gentoo.

Signed-off-by: Michał Górny <mgorny@gentoo.org>