net-misc/curl: Make it easier to disable QUIC
Until now we had a REQUIRED_USE that said "If you enable one of the QUIC
backends you must also enable USE=quic". This has required users that
wish to disable quic globally to also disable the quic backends on curl
in order for portage to successfully resolve a common package.

We're dropping this requirement; the remaining REQUIRED_USE are
sufficient to ensure that users who want QUIC have a backend enabled.

Closes: https://bugs.gentoo.org/956483
Signed-off-by: Matt Jolly <kangie@gentoo.org>