app-text/sigil: drop qtbase[cups] dependency
USE="gui widgets" provides qtprintsupport, *not* cups. Spare users from
potentially having to set it for nothing. Only a few rare packages
(typically using private cups headers) need this and it does not seem
to be the case here.

wrt USE=gui, it is implied by USE=widgets due to REQUIRED_USE, but feel
it's poor practice to rely on this so add it for printsupport.

Not really worth a revbump.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>