dev-python/pytest-bdd: skip test broken with current gherkin-official
Just noticed when running the test suite again, does not seem to break
the only revdep (qutebrowser) and would rather not restore and pin to
older gherkin-official, so just skip for now (unsure for the latest
version that worked but at least 30.0.4 was fine).

At same time cleanup unnecessary comments and keep a note for EAPI=9
which disables autoload by default when EPYTEST_PLUGINS is empty.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>