profiles/targets/desktop/plasma: enable opencv[features2d]
kdenlive->frei0r-plugins needs this, so enable it by default, at least
for now. It's a bit of a shame given slimmed-down opencv was partly
done for the benefit of Spectacle, but we can revisit later.

Right now, this achieves consistency given we set things for the benefit
of kdenlive and digikam.

Closes: https://bugs.gentoo.org/958961
Signed-off-by: Sam James <sam@gentoo.org>