dev-python/pytest-bdd: use EPYTEST_PLUGINS
Being empty replaces AUTOLOAD=1 as the Gentoo standard way to set this
now, albeit still need to set PYTEST_PLUGINS manually to avoid
depending on self in pytest-* ebuilds.

Formerly only kept a note about it given directly setting AUTOLOAD=1
does the same thing (with the idea to drop it in eapi 9), but
EPYTEST_PLUGINS made its way into the python guide and is added
by default in vim templates now, so may as well use it. Can optionally
drop it when empty in eapi 9 still.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>