sci-chemistry/gromacs: fixing gromacs opencl build for intel internal graphics
Intel integrated GPUs require:
    -DGMX_GPU_NB_CLUSTER_SIZE=4
to be set, otherwise running a GPU job will result in GPU not found.

I added the video_cards_intel use flag, in order to check if we build gromacs
for intel integrated graphics. If this is the case GMX_GPU_NB_CLUSTER_SIZE=4
will be set instead of the default 8.

With this patch I get about 3x performance on an Intel Core Ultra 7 256 machine,
utilizing integrated graphics that would be otherwise not found.

Signed-off-by: Thomas Haschka <haschka@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/42774
Closes: https://github.com/gentoo/gentoo/pull/42774
Signed-off-by: Sam James <sam@gentoo.org>