x11-misc/spnavcfg: fix build w/ qt-6.9 (with LTO, at least)
Workaround for bug #957962 (and bug #957446#c16). Building via
autotools requires a macro like bitcoin or gpgme used in the past
to ensure that -mno-direct-extern-access is used if Qt itself was
built with it.

The flag isn't in Qt's .pc files. It's easiest to just workaround this with -fPIC,
but in future, we might either have an eclass helper for this, or just encourage people
to use the m4 macro.

Bug: https://bugs.gentoo.org/957446
Closes: https://bugs.gentoo.org/957962
Signed-off-by: Sam James <sam@gentoo.org>