llvm-runtimes/libcxx: Support crosspkg build
Inherit the `crossdev` eclass, use clang for the current `CTARGET` and
install libcxx into a cross sysroot, if package belongs to any crossdev
category. That allows to bootstrap libcxx with crossdev by installing
`cross_*/libcxx`.

Extend the compiler test with the case when the (cross) toolchain has
a runtime, but doesn't have stdlib yet (because it's about to be built).
In such case, a build with default arguments fails, but a build with
`-nostdlib` succeeds. libc++ doesn't need stdlib to build, so that's
fine.

Signed-off-by: Michal Rostecki <vadorovsky@disroot.org>