games-fps/worldofpadman: add 1.7.0
- Bump to EAPI=8
- Build system changed to cmake
- Some USE flags were removed due to their corresponding build options being removed upstream.
- Dropped ~x86 keywords due to the above.

Vulkan support was also added upstream but I could not get the game to
run with it (segfault at launch).

This ebuild needs to pull both the "unified" zip from upstream releases
as well as the source tarball because the former contains the game
assets which are unfortunately stored in Git LFS and are not present in
the latter. Should that be worked around, this ebuild can be easily
adapted to pack the game assets into the required .pk3 files.

Signed-off-by: Kenton Groombridge <concord@gentoo.org>