games-emulation/pcsx2: update live
Had kind of feared this, but seem like patching for pluto(s)vg will
be annoying, so convert to a sed. The version requirement also does
not "seem" to matter and it'll avoid some headaches to remove it.

The tabs in the sed aren't stripped like bash's -heredocs, but it
does not hurt to have them for ebuild readability.

wrt vulkan, currently fails to build and upstream does not really
test disabling this. Could report/patch but would rather not deal with
this on this package even if it adds a decent deal to build time for
those that do not use vulkan (due to bundled glslang). May try again
eventually if upstream starts paying attention to it.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>