llvm-runtimes/libunwind: Support build with cross emerge wrapper
When cross emerge wrapper is used, using `llvm_prepend_path` without
any options, results in prepending the following path:

```
${ESYSROOT}/usr/lib/llvm/${slot}/bin
```

For example:

```
/usr/aarch64-unknown-linux-musl/usr/lib/llvm/19/bin
```

The problem is that cross emerge wrapper uses a cross clang wrapper as
a compiler, e.g. `aarch64-unknown-linux-musl-clang`. However, the file
with such name can be usually found in two places:

* `/usr/lib/llvm/19/bin/aarch64-unknown-linux-musl-clang`, which is
  a clang wrapper that calls the host clang with appropriate
  configuration. We want to call that file.
* `/usr/aarch64-unknown-linux-musl/usr/lib/llvm/19/bin/aarch64-unknown-linux-musl-clang`,
  which is an actual aarch64 compiler binary. We cannot execute it on
  x86_64, or other incompatible host.

However, the path added by `llvm_prepend_path` causes resolution of
`aarch64-unknown-linux-musl-clang` to the second path, therefore causing
errors like:

```
/usr/aarch64-unknown-linux-musl/usr/lib/llvm/19/bin/clang: cannot execute binary file: Exec format error
```

```
/usr/aarch64-unknown-linux-musl/usr/lib/llvm/19/bin/llvm-ar: cannot execute binary file: Exec format error
```

Solve that by using `llvm_prepend_path` with `-b` option, which uses
`BROOT` instead of `ESYSROOT`. That causes resolution to the first path.
The binaries in the slotted LLVM directory are used for building the
binary artifacts, so it's an intended usage of `-b`.

Signed-off-by: Michal Rostecki <vadorovsky@disroot.org>