dev-ruby/unf: fix 0.2.0 gemspec
Strip gem.extensions from the gemspec because we split out unf_ext. We
already have a sed to handle this but it needed adaptation to handle
upstream commit 04e206c2ee770c86889bc49d64cb57aa061ca6f5 in 0.2.0.

Without this, we get a load of "Ignoring unf-0.2.0 because its extensions are not built"
noise, but unf doesn't install an extension in Gentoo, we leave it to dev-ruby/unf_ext.

Signed-off-by: Sam James <sam@gentoo.org>