sys-kernel/vanilla-kernel: Use sha256sums.asc for verify-sig
Replace the unpacked tarball signature verification with verification
via signed sha256sums file.  While this is not the method recommended
upstream, it is pretty solid as well and it has the advantage
of covering all the files, including compressed tarballs and patches.
Unfortunately, the file is a moving target, so we need to rename it
and expect checksum mismatches when people skip Gentoo mirrors -- but
I guess that's okay for verify-sig.  Bumping SHA256SUM_DATE will be
taken care of by the bump-kernels script.

Signed-off-by: Michał Górny <mgorny@gentoo.org>