sys-kernel/vanilla-kernel: Use patch-based reconstruction for 6.15.4
Rather than using the complete versioned tarball, use the first release
archive along with upstream patch file to reconstruct specific version.
This should significantly reduce bandwidth use for updates, and Gentoo
mirror load -- as for every patch bump we will be using a patch
of up to a few MiB rather than a new ~150 MiB archive.  This is also
part of an effort to streamline vanilla-kernel and gentoo-kernel
ebuilds.

Signed-off-by: Michał Górny <mgorny@gentoo.org>