media-sound/easytag: drop confusing dependency on libcanberra-gtk
This was added in commit 9e0426f0b0345675926e1635d440f429a0ec1fad with
zero explanation as part of a routine version bump. The package doesn't
look for or use it, there is zero mention of canberra at all ANYWHERE in
git history, and other distros don't depend on it. I daresay it should
never have been added.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>