dot-a.eclass: make strip-lto remember if guarantee-fat ran
The current implementation checks repeatedly if `tc-is-lto` is true.  In
addition to running a small compile check redundantly, this has a
potential correctness downside. An ebuild could do this:

```
lto-guarantee-fat
tc-is-lto && myeconfargs+=( --enable-lto )
filter-lto
```

The idea is, the configure script has special handling and simply adding
-flto is too "simple" to work. e.g. not all sources build with LTO.

When this happens, strip-lto-bytecode "thinks" we didn't build with LTO
at all, and returns early. By tracking state across functions -- and
across phases -- we can reliably tell if `lto-guarantee-fat` actually
"succeeded".

Bug: https://bugs.gentoo.org/958412
Acked-by: Sam James <sam@gentoo.org>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>