profiles/features/musl: add sys-apps/musl-locales to @system
musl by default does not provide any sort of locale support, and so does
not have a `locale` binary by default. This often results in spurious
command not found QA notices or in extreme cases even build failures.
Adelie's musl-locales project has been adopted as the de facto solution
for locales on musl, so having it in @system for consistency with glibc
is probably a good idea.

Signed-off-by: Violet Purcell <vimproved@inventati.org>