dev-lang/python: Update dev-python/ensurepip-* dependencies
Inline dev-python/ensurepip-wheels dependencies in all Python versions
to pave the way for removing the metapackage.  Remove unnecessary
dev-python/ensurepip-wheel test dependency from slots other than 3.12;
the dependency was introduced in that version and removed in 3.13,
as the bundled setuptools wheel was updated.

TODO: patch 3.12 to remove the dependency as well?
Signed-off-by: Michał Górny <mgorny@gentoo.org>